+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Release 5.12.0 (unreleased):
  - ISO mdoc:
    - Preserve `Document.errors` in parsed ISO document results instead of failing validation
  - Change: Executing unsatisfiable DCQL queries no longer throws on matching, only on submission.
+ - Change: Update DCQLClaimsQuery to OpenID4VP 1.0
 
 Release 5.11.0:
  - Digital Credentials API:
