+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/dcql/DCQLSdJwtCredentialMetadataAndValidityConstraints.kt
@@ -22,9 +22,9 @@ data class DCQLSdJwtCredentialMetadataAndValidityConstraints(
         const val VCT_VALUES = "vct_values"
     }
 
-    fun validate(actualCredentialType: String?): KmmResult<Unit> = catching {
-        if (actualCredentialType !in vctValues) {
-            throw IllegalArgumentException("Incompatible SD-JWT document type")
+    fun validate(actualCredentialType: String): KmmResult<Unit> = catching {
+        require(actualCredentialType in vctValues) {
+            "Incompatible SD-JWT document type"
         }
     }
 }
